Add files via upload
User Authentication with Tkinter box. In case of an incorrect user has been entered the program will open the local camera and will screenshot the screen. The photo will be saved on the desktop but can be redirected to the desired cloud folder. After this message notifying the user will be displayed and the program will end.